requirements.txt
flask
python-whois